refactor: move all stories inside __stories__ directories
