+++ .storybook/main.ts
@@ -5,10 +5,11 @@ import { basename, resolve } from 'path';
 import TsconfigPathsPlugin from 'tsconfig-paths-webpack-plugin';
 
 function findStories(): string[] {
-  const stories = glob.sync('../src/{components,stories}/**/*.stories.tsx', {
+  const stories = glob.sync('../src/**/__stories__/*.stories.tsx', {
     cwd: __dirname,
   });
 
+  // Will order stories alphabetically by file name. If a title is specified it will take priority over the file name.
   const alphabetical = stories.sort((a: string, b: string) => {
     const storyNameA = getTitle(resolve(__dirname, a)) || a;
     const storyNameB = getTitle(resolve(__dirname, b)) || b;
@@ -19,6 +20,7 @@ function findStories(): string[] {
   return alphabetical;
 }
 
+// Read the specified title from a story meta.
 function getTitle(path: string): string | null {
   const contents = readFileSync(path, { encoding: 'utf-8' });
   const match = contents.match(/const \w+: Meta = {.+\stitle: '(.+?)',/s);
