+++ src/components/dropzone/__stories__/Dropzone.stories.tsx
