+++ src/components/input/__stories__/SearchInput.stories.tsx
