+++ src/components/slider/Slider.tsx
