+++ src/components/slider/__stories__/Slider.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
-import { Slider, SliderProps } from 'src/components/radix/Slider';
+import { Slider, SliderProps } from 'src/components/slider/Slider';
 
 const SliderMeta: Meta = {
   component: Slider,
