+++ src/components/styles/__stories__/Colors.stories.tsx
