+++ src/components/text-avatar/__stories__/TextAvatar.stories.tsx
