--- src/stories/ScratchPad/ScratchPad.stories.tsx
@@ -1,10 +1,0 @@
-import { Meta } from '@storybook/react/types-6-0';
-
-import { ScratchPad } from './ScratchPad';
-
-const ScratchPadMeta: Meta = {
-  component: ScratchPad,
-};
-
-export default ScratchPadMeta;
-export const ScratchPadBasic = ScratchPad;
