+++ tsconfig.emitDeclarationOnly.json
@@ -6,5 +6,5 @@
     "declarationDir": "dist"
   },
   "include": ["./src"],
-  "exclude": ["**/__tests__", "**/__stories__", "**/*.stories.*"]
+  "exclude": ["**/_*"]
 }
